third_party/WebKit/LayoutTests/media/video-force-preload-none.html

<!doctype html>
<title>Tests that the document setting for forcing preload none works.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video preload="metadata"></video>
<script>
async_test(function(t) {
  internals.settings.setForcePreloadNoneForMediaElements(true);

  var video = document.querySelector('video');
  assert_equals(video.preload, 'none');
  video.src = 'resources/test-positive-start-time.webm';

  video.onsuspend = t.step_func_done();
  video.onprogress = t.unreached_func();
  t.add_cleanup(function() {
    internals.settings.setForcePreloadNoneForMediaElements(false);
  });
});
</script>
